refactor(AddButton): 가독성 향상을 위한 컴포넌트화
