+++ src/components/ProductItem/index.tsx
@@ -1,12 +1,11 @@
 import * as S from './style';
 
-import { useContext, useEffect, useState } from 'react';
+import { useContext } from 'react';
 
 import { UseCartItemsContext } from '../../App';
 import AdjustButtonContainer from '../AdjustButtonContainer';
-import Spinner from '../Spinner';
 
-import { ADD_TO_CART } from '../../assets/images';
+import AddButton from '../AddButton';
 
 interface ProductItemProps {
   id: number;
@@ -16,27 +15,12 @@ interface ProductItemProps {
 }
 
 const ProductItem = ({ id, imageUrl, name, price }: ProductItemProps) => {
-  const { getCartItems, addCartItem } = useContext(UseCartItemsContext);
-  const [isLoading, setIsLoading] = useState(false);
+  const { getCartItems } = useContext(UseCartItemsContext);
 
   const cartItem = getCartItems.data
     ? getCartItems.data.find(({ product }) => id === product.id)
     : undefined;
   const isInCart = !!cartItem;
-
-  const BUTTON_TEXT = '담기';
-
-  const handleAddToCart = () => {
-    setIsLoading(true);
-
-    addCartItem.mutate(id, {
-      onError: () => setIsLoading(false),
-    });
-  };
-
-  useEffect(() => {
-    setIsLoading(false);
-  }, [cartItem?.quantity]);
 
   return (
     <S.ProductItem>
@@ -47,17 +31,7 @@ const ProductItem = ({ id, imageUrl, name, price }: ProductItemProps) => {
           <S.Price>{price.toLocaleString('ko-KR')}원</S.Price>
         </S.Information>
         <S.ButtonContainer>
-          {!isInCart && (
-            <S.AddButton onClick={handleAddToCart} $isInCart={isInCart}>
-              {!isLoading && (
-                <>
-                  <S.ButtonImage src={ADD_TO_CART} />
-                  <span>{BUTTON_TEXT}</span>
-                </>
-              )}
-              {isLoading && <Spinner />}
-            </S.AddButton>
-          )}
+          {!isInCart && <AddButton id={id} />}
           {isInCart && <AdjustButtonContainer cartItem={cartItem} />}
         </S.ButtonContainer>
       </S.InformationContainer>
